easy/21/merge.go
package easy

/*

Merge two lists in a one sorted list.

The list should be made by splicing together the nodes of the first two
lists.




*/

// & Helper functions / types

type ListNode struct {
	Val  int
	Next *ListNode
}

func createLinkedList(arr []int) *ListNode {
	if len(arr) == 0 {
		return nil
	}

	head := &ListNode{Val: arr[0]}
	current := head
	for i := 1; i < len(arr); i++ {
		current.Next = &ListNode{Val: arr[i]}
		current = current.Next
	}
	return head
}

func linkedListToArr(head *ListNode) []int {
	arr := make([]int, 0)
	current := head

	for current != nil {
		arr = append(arr, current.Val)
		current = current.Next
	}
	return arr
}

// & Working solution

func mergeTwoLists(l1 *ListNode, l2 *ListNode) *ListNode {
	dummy := &ListNode{}

	current := dummy

	for l1 != nil && l2 != nil {
		if l1.Val <= l2.Val {
			current.Next = l1
			l1 = l1.Next
		} else {
			current.Next = l2
			l2 = l2.Next
		}
		current = current.Next

	}

	if l1 != nil {
		current.Next = l1
	} else {
		current.Next = l2
	}

	return dummy.Next
}
